More tests, finished porting
