--- specs/Routes.spec.js
@@ -1,60 +1,0 @@
-require('./helper');
-var Route = require('../modules/components/Route');
-var Routes = require('../modules/components/Routes');
-
-describe('a Routes', function () {
-
-  describe('when a transition is aborted', function () {
-    it('triggers onAbortedTransition', function (done) {
-      var App = React.createClass({
-        statics: {
-          willTransitionTo: function (transition) {
-            transition.abort();
-          }
-        },
-        render: function () {
-          return React.DOM.div();
-        }
-      });
-
-      function handleAbortedTransition(transition) {
-        assert(transition);
-        done();
-      }
-
-      var routes = ReactTestUtils.renderIntoDocument(
-        Routes({ onAbortedTransition: handleAbortedTransition },
-          Route({ handler: App })
-        )
-      );
-    });
-  });
-
-  describe('when there is an error in a transition hook', function () {
-    it('triggers onTransitionError', function (done) {
-      var App = React.createClass({
-        statics: {
-          willTransitionTo: function (transition) {
-            throw new Error('boom!');
-          }
-        },
-        render: function () {
-          return React.DOM.div();
-        }
-      });
-
-      function handleTransitionError(error) {
-        assert(error);
-        expect(error.message).toEqual('boom!');
-        done();
-      }
-
-      var routes = ReactTestUtils.renderIntoDocument(
-        Routes({ onTransitionError: handleTransitionError },
-          Route({ handler: App })
-        )
-      );
-    });
-  });
-
-});
