--- specs/main.js
@@ -1,9 +1,0 @@
-require('./ActiveDelegate.spec.js');
-require('./AsyncState.spec.js');
-require('./DefaultRoute.spec.js');
-require('./NotFoundRoute.spec.js');
-require('./Path.spec.js');
-require('./PathStore.spec.js');
-require('./Route.spec.js');
-require('./Routes.spec.js');
-require('./RouteStore.spec.js');
